+++ OpenXRechnungToolboxMac.sh
@@ -1,1 +1,1 @@
-java -XstartOnFirstThread -Dlog4j.configuration=./resources/log4j2.xml -jar OpenXRechnungToolbox.jar "resources/app.config"
+java -XstartOnFirstThread -Dlog4j.configuration=./resources/log4j2.xml -jar OpenXRechnungToolbox.jar --config resources\app.config