+++ docs/javadoc/package-search-index.js
@@ -1,1 +1,0 @@
-packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"org.oxt.toolbox.converter"},{"l":"org.oxt.toolbox.gui"},{"l":"org.oxt.toolbox.helpers"},{"l":"org.oxt.toolbox.leitwegcheck"},{"l":"org.oxt.toolbox.validation"},{"l":"org.oxt.toolbox.visualization"}];updateSearchResults();