+++ src/main/java/org/oxt/toolbox/gui/AppWindow.java
@@ -7,7 +7,12 @@ import java.io.StringWriter;
 import java.net.MalformedURLException;
 import java.util.ResourceBundle;
 
-import org.apache.logging.log4j.LogManager;
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.HelpFormatter;
+import org.apache.commons.cli.Option;
+import org.apache.commons.cli.OptionGroup;
+import org.apache.commons.cli.Options;
+import org.apache.commons.cli.ParseException;
 import org.apache.logging.log4j.Logger;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.dnd.DND;
@@ -25,6 +30,9 @@ import org.eclipse.swt.widgets.Menu;
 import org.eclipse.swt.widgets.MenuItem;
 import org.eclipse.swt.widgets.MessageBox;
 import org.eclipse.swt.widgets.Shell;
+import org.oxt.toolbox.cli.CLICii2ubl;
+import org.oxt.toolbox.cli.CLIValidation;
+import org.oxt.toolbox.cli.CLIVisualization;
 import org.oxt.toolbox.converter.ConverterImpl;
 import org.oxt.toolbox.helpers.AppProperties;
 import org.oxt.toolbox.helpers.CustomDropTargetAdapter;
@@ -39,7 +47,7 @@ import org.oxt.toolbox.visualization.VisualizerImpl;
  * Main class of OpenXRechnungToolbox.
  * Builds the main GUI window. 
  * @author Dr. Jan C. Thiele
- * @version 1.1.2
+ * @version 3.0.0
  */
 public class AppWindow {
 
@@ -123,11 +131,10 @@ public class AppWindow {
                 	  VisualizerImpl viz = new VisualizerImpl();
                 	  File invoiceFile = new File(base,name);
                 	  try {
-                        StringWriter sw = viz.runVisualization(name, invoiceFile.getAbsolutePath(), AppProperties.prop.getProperty("viz.intermediate.ubl.xsl"), AppProperties.prop.getProperty("viz.intermediate.ublcn.xsl"), AppProperties.prop.getProperty("viz.intermediate.cii.xsl"), AppProperties.prop.getProperty("viz.html.xsl"));
+                        StringWriter sw = viz.runVisualization(/*name,*/ invoiceFile.getAbsolutePath(), AppProperties.prop.getProperty("viz.intermediate.ubl.xsl"), AppProperties.prop.getProperty("viz.intermediate.ublcn.xsl"), AppProperties.prop.getProperty("viz.intermediate.cii.xsl"), AppProperties.prop.getProperty("viz.html.xsl"));
                         win.launch(display, viz, null, name, invoiceFile.getAbsolutePath(), sw.toString());
 	                  } catch (Exception e1) {
 	                	  logger.error(e1.getMessage());
-	                	  //e1.ackTrace();
 	                      String html = resourceBundle.getString("vizError");
 	                      win.launch(display, viz, null, name, invoiceFile.getAbsolutePath(), html);	      				
 	                  }
@@ -149,13 +156,10 @@ public class AppWindow {
                 	  ValidatorImpl vali = new ValidatorImpl();
                 	  File invoiceFile = new File(base,name);
                 	  try {
-                        //StringWriter sw = vali.runValidation(name, invoiceFile.getAbsolutePath(), AppProperties.prop.getProperty("validator.scenario"));
-                        //win.launch(display, null, vali, name, invoiceFile.getAbsolutePath(), sw.toString());
-                        String sw = vali.runValidation(name, invoiceFile.getAbsolutePath(), AppProperties.prop.getProperty("validator.scenario"));
+                        String sw = vali.runValidation(invoiceFile.getAbsolutePath(), AppProperties.prop.getProperty("valiVersion"));
                         win.launch(display, null, vali, name, invoiceFile.getAbsolutePath(), sw);
 	                  } catch (Exception e1) {
 	                	  logger.error(e1.getMessage());
-	                	  //e1.ackTrace();
 	                      String html = resourceBundle.getString("valiError");	                      
 	                      win.launch(display, null, vali, name, invoiceFile.getAbsolutePath(), html);	      				
 	                  }
@@ -330,27 +334,130 @@ public class AppWindow {
 	/**
 	 * Entry point.
 	 * @param args command line arguments, first argument used for config file path
-	 * @throws IOException error when config and/or language files not found
+	 * @throws Exception 
 	 */
-	public static void main(String[] args) throws IOException {		
+	public static void main(String[] args) throws Exception {		
 		System.setProperty("com.sun.xml.bind.v2.bytecode.ClassTailor.noOptimize","true");
 		
 		String config = "resources/app.config";
-		if (args.length > 0) {
-			config = args[0];
-		}
+
+		// Collection of arguments for testing
+		
+		//--config resources/app2.config
+
+		// -val -i exampleInvoices\XRechnung_v3.0.2\01.01a-INVOICE_ubl.xml -o testreport1.html -v 3.0.2
 		
 		
-		AppProperties.initializeProperties(config);
-		if (AppProperties.prop == null) {
-			Display display = new Display();
-			Shell shell = new Shell(display);
-			MessageBox messageBox = new MessageBox(shell, SWT.ICON_WARNING | SWT.OK);				        
-	        messageBox.setText("Error");
-	        messageBox.setMessage("Properties file not found.");
-	        messageBox.open();
-		}
-		new AppWindow();
+		// -viz -i exampleInvoices\XRechnung_v3.0.2\01.01a-INVOICE_ubl.xml -o tesviz.html 
+		
+		// -viz -i exampleInvoices\XRechnung_v3.0.2\01.01a-INVOICE_ubl.xml -o tesviz.html 
+		// -viz -i exampleInvoices\XRechnung_v3.0.2\01.01a-INVOICE_ubl.xml -o tesviz.pdf -p
+		
+		// -con -i exampleInvoices\XRechnung_v3.0.2\01.01a-INVOICE_uncefact.xml -o testconUBL.xml
+		
+		Options baseOptions = new Options();
+		
+		OptionGroup vizvalGroup = new OptionGroup();
+        Option vizOption = new Option("viz", "visualization", false, "run visualization without GUI in Command Line Interface (CLI) mode.");
+        vizOption.setRequired(false);
+        vizvalGroup.addOption(vizOption);
+        Option valOption = new Option("val", "validation", false, "run validation without GUI in Command Line Interface (CLI) mode.");
+        valOption.setRequired(false);
+        vizvalGroup.addOption(valOption);
+        Option convOption = new Option("con", "conversion", false, "run conversion from CII to UBL without GUI in Command Line Interface (CLI) mode.");
+        convOption.setRequired(false);
+        vizvalGroup.addOption(convOption);
+        baseOptions.addOptionGroup(vizvalGroup);
+
+        Option vizInputOption = new Option("i", "input", true, "input invoice file path for cli.");
+        vizInputOption.setRequired(false);
+        baseOptions.addOption(vizInputOption);
+        Option vizOutputOption = new Option("o", "output", true, "output file path for cli (either validation report or visualization file).");
+        vizOutputOption.setRequired(false);
+        baseOptions.addOption(vizOutputOption);
+        Option vizPdfOption = new Option("p", "pdf", false, "visualization as pdf (instead of html) for cli visualization.");
+        vizPdfOption.setRequired(false);
+        baseOptions.addOption(vizPdfOption);                
+        Option valValiVersionOption = new Option("v", "valiversion", true, "version of XRechnung to validate against for cli validation, e.g. 3.0.2. Must be in list of version of config file.");
+        valValiVersionOption.setRequired(false);
+        baseOptions.addOption(valValiVersionOption);
+        Option configOption = new Option("c", "config", true, "config file path.");
+        configOption.setRequired(false);
+        baseOptions.addOption(configOption);
+
+        org.apache.commons.cli.CommandLineParser parser = new org.apache.commons.cli.DefaultParser();
+        HelpFormatter help = new HelpFormatter();
+
+        try {
+        	CommandLine cmd = parser.parse(baseOptions, args);
+
+        	if (cmd.hasOption("config")) {
+	            String alternativeConfig = cmd.getOptionValue("config");
+	            if (alternativeConfig != null) {
+	            	if (alternativeConfig.length() > 0) {
+	                	File configFile = new File(alternativeConfig);
+	                	if (configFile.exists()) {
+	                		config = alternativeConfig;
+	                	} else {
+	                		throw new ParseException("config file not found at: "+alternativeConfig);
+	                	}
+	                }
+	            }  
+        	}
+        	
+        	if (cmd.hasOption("visualization") || cmd.hasOption("validation") || cmd.hasOption("conversion")) {
+	        	// CLI mode
+        		if (cmd.hasOption("visualization")) {
+            		if (cmd.hasOption("input") & cmd.hasOption("output")) {
+            			CLIVisualization visualization = new CLIVisualization();
+            			visualization.cliVisualization(cmd.getOptionValue("input"), cmd.getOptionValue("output"), cmd.hasOption("pdf"), config);
+            		}
+            		else {
+            			throw new ParseException("not enough arguments for cli visualition");
+            		}
+        		}
+        		if (cmd.hasOption("validation")) {
+            		if (cmd.hasOption("input") & cmd.hasOption("output") & cmd.hasOption("valiversion")) {
+            			CLIValidation validation = new CLIValidation();
+            			validation.cliValidation(cmd.getOptionValue("input"), cmd.getOptionValue("output"), cmd.getOptionValue("valiversion"), config);            			
+            		}
+            		else {
+            			throw new ParseException("not enough arguments for cli validation");
+            		}            		
+        		}
+        		if (cmd.hasOption("conversion")) {
+        			if (cmd.hasOption("input") & cmd.hasOption("output")) {
+        				CLICii2ubl conversion = new CLICii2ubl();
+        				conversion.cliCii2ubl(cmd.getOptionValue("input"), cmd.getOptionValue("output"), config);
+        			}
+            		else {
+            			throw new ParseException("not enough arguments for cli conversion");
+            		}          			
+        		}
+	        } 
+	        else {
+	        	// GUI mode
+	    		AppProperties.initializeProperties(config);
+	    		if (AppProperties.prop == null) {
+	    			Display display = new Display();
+	    			Shell shell = new Shell(display);
+	    			MessageBox messageBox = new MessageBox(shell, SWT.ICON_WARNING | SWT.OK);				        
+	    	        messageBox.setText("Error");
+	    	        messageBox.setMessage("Properties file not found.");
+	    	        messageBox.open();
+	    		}
+	    		
+	    		new AppWindow();	        	
+	        }
+        } catch (ParseException e) {
+	        System.out.println(e.getMessage());
+	        help.printHelp("\n"
+	        		+ "Gui mode: java -Dlog4j.configuration=./resources/log4j2.xml -jar OpenXRechnungToolbox.jar --config <path to config file> \n"
+	        		+ "CLI mode validation: java -Dlog4j.configuration=./resources/log4j2.xml -jar OpenXRechnungToolbox.jar --validation --input <path to invoice file> --output <path of report file> --valiVersion <version of XRechnung to validate against> \n"
+	        		+ "CLI mode visualization: java -Dlog4j.configuration=./resources/log4j2.xml -jar OpenXRechnungToolbox.jar --visualization --input <path to invoice file> --output <path of visualization file> --type <pdf or html>", baseOptions);
+	        System.exit(1);
+	    }
+        				
 	}
 
 }